entrypoint update

+++ entrypoint.client.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
-npm run build && mkdir -p /app/html && cp -rf /app/build/* /app/html &
+if [ ! -e /app/buildhasrun ]
+    npm run build && mkdir -p /app/html && cp -rf /app/build/* /app/html &
+    touch /app/buildhasrun
+fi
+
 npm run start:prod
